split HEI into two collections of functions
